refactor: separate media from summary
