+++ projects/client/src/lib/sections/summary/EpisodeSummary.svelte
@@ -10,13 +10,13 @@
   import MarkAsWatchedAction from "$lib/sections/media-actions/mark-as-watched/MarkAsWatchedAction.svelte";
   import { UrlBuilder } from "$lib/utils/url/UrlBuilder";
   import type { EpisodeSummaryProps } from "./components/EpisodeSummaryProps";
-  import MediaMetaInfo from "./components/MediaMetaInfo.svelte";
-  import MediaOverview from "./components/MediaOverview.svelte";
-  import MediaSummaryActions from "./components/MediaSummaryActions.svelte";
-  import MediaSummaryContainer from "./components/MediaSummaryContainer.svelte";
-  import MediaSummaryHeader from "./components/MediaSummaryHeader.svelte";
-  import MediaTitle from "./components/MediaTitle.svelte";
+  import MediaMetaInfo from "./components/media/MediaMetaInfo.svelte";
   import RateNowButton from "./components/rating/RateNowButton.svelte";
+  import SummaryActions from "./components/summary/SummaryActions.svelte";
+  import SummaryContainer from "./components/summary/SummaryContainer.svelte";
+  import SummaryHeader from "./components/summary/SummaryHeader.svelte";
+  import SummaryOverview from "./components/summary/SummaryOverview.svelte";
+  import SummaryTitle from "./components/summary/SummaryTitle.svelte";
 
   const {
     episode,
@@ -48,7 +48,7 @@
 
 <CoverImageSetter src={episode.cover.url ?? ""} type="show" />
 
-<MediaSummaryContainer>
+<SummaryContainer>
   {#snippet poster()}
     <SummaryPoster
       href={UrlBuilder.show(show.slug)}
@@ -63,14 +63,14 @@
     </SummaryPoster>
   {/snippet}
 
-  <MediaSummaryHeader {title}>
-    <MediaTitle {title} />
+  <SummaryHeader {title}>
+    <SummaryTitle {title} />
     <Link href={UrlBuilder.show(show.slug)}>
       <h6>{showTitle}</h6>
     </Link>
     <p class="meta-info">{m.season_episode_number_label(episode)}</p>
     <GenreList genres={show.genres} />
-  </MediaSummaryHeader>
+  </SummaryHeader>
 
   <MediaMetaInfo
     certification={show.certification}
@@ -81,17 +81,17 @@
     {watchers}
   />
 
-  <MediaOverview {title} {overview} />
+  <SummaryOverview {title} {overview} />
 
   <RenderFor audience="authenticated">
-    <MediaSummaryActions>
+    <SummaryActions>
       <RateNowButton type="episode" id={episode.id} />
 
       <RenderFor device={["mobile", "tablet-sm"]} audience="authenticated">
         {@render mediaActions()}
       </RenderFor>
-    </MediaSummaryActions>
+    </SummaryActions>
   </RenderFor>
-</MediaSummaryContainer>
+</SummaryContainer>
 
 <SeasonList {show} {seasons} />
