+++ projects/client/src/lib/sections/summary/ShowSummary.svelte
@@ -12,8 +12,8 @@
   import CastList from "../lists/CastList.svelte";
   import RelatedList from "../lists/RelatedList.svelte";
   import SeasonList from "../lists/SeasonList.svelte";
-  import MediaSummary from "./components/MediaSummary.svelte";
-  import type { MediaSummaryProps } from "./components/MediaSummaryProps";
+  import MediaSummary from "./components/media/MediaSummary.svelte";
+  import type { MediaSummaryProps } from "./components/media/MediaSummaryProps";
 
   type ShowSummaryProps = MediaSummaryProps<ShowSummary> & {
     stats: MediaStats;
