--- projects/client/src/lib/sections/summary/components/MediaSummary.ts
@@ -1,4 +1,0 @@
-import type { MovieSummary } from '$lib/requests/models/MovieSummary.ts';
-import type { ShowSummary } from '$lib/requests/models/ShowSummary.ts';
-
-export type MediaSummary = MovieSummary | ShowSummary;
