+++ projects/client/src/lib/sections/summary/components/media/MediaCollapsableValues.svelte
