+++ projects/client/src/lib/sections/summary/components/media/MediaSummaryProps.ts
